refactor: domain model package
