+++ app/src/main/java/com/voyager/domain/model/ride/address/Address.kt
@@ -1,4 +1,4 @@
-package com.voyager.domain.ride.address
+package com.voyager.domain.model.ride.address
 
 import kotlinx.serialization.SerialName
