+++ app/src/main/java/com/voyager/domain/model/ride/estimate/Option.kt
@@ -1,4 +1,4 @@
-package com.voyager.domain.ride.estimate
+package com.voyager.domain.model.ride.estimate
 
 data class Option (
     val id: Int,
