+++ app/src/main/java/com/voyager/domain/model/ride/estimate/RideEstimateError.kt
@@ -1,6 +1,6 @@
 // YApi QuickType插件生成，具体参考文档:https://plugins.jetbrains.com/plugin/18847-yapi-quicktype/documentation
 
-package com.voyager.domain.ride.estimate
+package com.voyager.domain.model.ride.estimate
 
 data class RideEstimateError (
     val errorDescription: String,
